Update time.js
in the first line, there seems to be a type that is globals instead of global. All other javascript files are imported to the html file, except those that have "globals". Could this be the reason why they are failing ? kindly pull this and send me the updated file.Thanks